Put pages into folders for future css files
