+++ client/src/pages/StartPage/StartPage.jsx
